% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agolExtract.R
\name{agolExtract}
\alias{agolExtract}
\title{Extract ArcGIS Online (AGOL) Web Feature Layers to Postgres and/or Geopackage}
\usage{
agolExtract(
  input_file,
  agol_portal,
  pg_db,
  pg_host,
  pg_port,
  pg_user,
  pg_password,
  email_to
)
}
\arguments{
\item{input_file}{string: input csv file defining the AGOL web layers to extract}

\item{agol_portal}{string: address of the AGOL portal containing the web layers to extract}

\item{pg_db}{string:  name of the destination Postgres database (if applicable)}

\item{pg_host}{string:  host of the destination Postgres database (if applicable)}

\item{pg_port}{integer:  port of the destination Postgres database (if applicable)}

\item{pg_user}{string:  user of the destination Postgres database (if applicable)}

\item{pg_password}{string:  password of the destination Postgres database (if applicable)}

\item{email_to}{string: email address to send to, if blank/NA then no email sent}

\item{resultant_table}{name of the existing resultant table.}
}
\value{
nothing is returned
}
\description{
Extract ArcGIS Online (AGOL) Web Feature Layers to Postgres and/or Geopackage
}
\examples{
coming soon
}
