% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentFileLocation.R
\name{getCurrentFileLocation}
\alias{getCurrentFileLocation}
\title{Get the Location of the currently running file}
\usage{
getCurrentFileLocation()
}
\value{
directory path of currently running R script
}
\description{
Get the Location of the currently running file
}
\examples{
coming soon

}
