% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOutlookRunning.R
\name{isOutlookRunning}
\alias{isOutlookRunning}
\title{Deterimine if Outlook Applicatinon is currently running}
\usage{
isOutlookRunning()
}
\value{
BOOLEAN (TRUE/FALSE)
}
\description{
Deterimine if Outlook Applicatinon is currently running
}
\examples{
coming soon
}
