% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{sendEmail}
\alias{sendEmail}
\title{Sends an email from the Outlook Application on the machine using the address of that logged in user}
\usage{
sendEmail(to, subject, body)
}
\arguments{
\item{to}{string: email address to send the email to}

\item{subject}{string: subject line of the email}

\item{body}{string: email body text content}
}
\value{
nothing is returned
}
\description{
Sends an email from the Outlook Application on the machine using the address of that logged in user
}
\examples{
coming soon

}
