% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDelta.R
\name{timeDelta}
\alias{timeDelta}
\title{Get the difference in two times}
\usage{
timeDelta(time1, time2)
}
\arguments{
\item{time1}{as.POSIXct(now()) formatted time value.  start time}

\item{time2}{as.POSIXct(now()) formatted time value.  end}
}
\value{
string: formatted string expressing time difference in decimal seconds, minutes or hours
}
\description{
Get the difference in two times
}
\examples{
coming soon

}
